# docs/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "迷因典 API 文檔",
    "version": "1.0.0",
    "description": "迷因典後端API的完整文檔，包含所有端點的詳細說明",
    "contact": {
      "name": "迷因典開發團隊",
      "email": "support@memedam.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "開發環境"
    },
    {
      "url": "https://api.memedam.com",
      "description": "生產環境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token認證"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "錯誤訊息"
          },
          "status": {
            "type": "integer",
            "description": "HTTP狀態碼"
          }
        }
      },
...
      "CreateAnnouncementRequest": {
        "type": "object",
        "required": ["title", "content"],
        "properties": {
          "title": {
            "type": "string",
            "description": "公告標題"
          },
          "content": {
            "type": "string",
            "description": "公告內容"
          },
          "priority": {
            "type": "integer",
            "default": 0,
            "description": "優先級"
          },
          "isActive": {
            "type": "boolean",
            "default": true,
            "description": "是否啟用"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "開始顯示時間"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "結束顯示時間"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/announcements": {
      "post": {
        "summary": "建立公告",
        "tags": ["Announcements"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "公告創建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "announcement": {
                      "$ref": "#/components/schemas/Announcement"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "get": {
        "summary": "取得所有公告",
        "tags": ["Announcements"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "頁碼"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每頁數量"
          },
          {
            "in": "query",
            "name": "active",
            "schema": {
              "type": "boolean"
            },
            "description": "只顯示啟用的公告"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "integer"
            },
            "description": "篩選優先級"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得公告列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "announcements": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Announcement"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/collections": {
      "post": {
        "summary": "收藏迷因",
        "tags": ["Collections"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "收藏成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "delete": {
        "summary": "取消收藏迷因",
        "tags": ["Collections"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "meme_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "迷因ID"
          }
        ],
        "responses": {
          "200": {
            "description": "取消收藏成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "get": {
        "summary": "取得用戶的收藏列表",
        "tags": ["Collections"],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "string"
            },
            "description": "用戶ID（可選，不提供則取得當前用戶）"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "頁碼"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每頁數量"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得收藏列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "collections": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Collection"
                      }
                    },
                    "memes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Meme"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          }
        }
      }
    },
...
    "/api/comments": {
      "post": {
        "summary": "建立新留言",
        "tags": ["Comments"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "留言創建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "comment": {
                      "$ref": "#/components/schemas/Comment"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "get": {
        "summary": "取得留言列表",
        "tags": ["Comments"],
        "parameters": [
          {
            "in": "query",
            "name": "meme_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "迷因ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "頁碼"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每頁數量"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": ["newest", "oldest", "most_liked"],
              "default": "newest"
            },
            "description": "排序方式"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得留言列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          }
        }
      }
    },
...
    "/api/dislikes": {
      "post": {
        "summary": "為迷因按噓",
        "tags": ["Dislikes"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDislikeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "按噓成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DislikeResponse"
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "delete": {
        "summary": "取消對迷因的噓",
        "tags": ["Dislikes"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "meme_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "迷因ID"
          }
        ],
        "responses": {
          "200": {
            "description": "取消噓成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "get": {
        "summary": "取得迷因的噓數和噓的用戶列表",
        "tags": ["Dislikes"],
        "parameters": [
          {
            "in": "query",
            "name": "meme_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "迷因ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "頁碼"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每頁數量"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得噓的資訊",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalDislikes": {
                      "type": "integer",
                      "description": "總噓數"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          }
        }
      }
    },
...
    "/api/likes": {
      "post": {
        "summary": "為迷因按讚",
        "tags": ["Likes"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLikeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "按讚成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeResponse"
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "delete": {
        "summary": "取消對迷因的讚",
        "tags": ["Likes"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "meme_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "迷因ID"
          }
        ],
        "responses": {
          "200": {
            "description": "取消讚成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "get": {
        "summary": "取得迷因的讚數和讚的用戶列表",
        "tags": ["Likes"],
        "parameters": [
          {
            "in": "query",
            "name": "meme_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "迷因ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "頁碼"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每頁數量"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得讚的資訊",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalLikes": {
                      "type": "integer",
                      "description": "總讚數"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          }
        }
      }
    },
...
    "/api/memes": {
      "post": {
        "summary": "建立新迷因",
        "tags": ["Memes"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "迷因創建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meme"
                }
              }
            }
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未授權"
          }
        }
      },
      "get": {
        "summary": "取得所有迷因",
        "tags": ["Memes"],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "頁碼"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "每頁數量"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": ["newest", "oldest", "popular", "hot"],
              "default": "newest"
            },
            "description": "排序方式"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得迷因列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Meme"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/recommendations": {
      "get": {
        "summary": "取得綜合推薦",
        "tags": ["Recommendations"],
        "parameters": [
          {
            "in": "query",
            "name": "algorithm",
            "schema": {
              "type": "string",
              "enum": [
                "hot",
                "latest",
                "mixed",
                "user-interest",
                "content-based",
                "tag-based",
                "collaborative-filtering",
                "social-collaborative-filtering"
              ],
              "default": "mixed"
            },
            "description": "推薦演算法"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "推薦數量限制"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得推薦",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}